refactored human detector class
